+++ lib/main.dart
@@ -13,7 +13,7 @@ import 'testpage.dart';
 final List<Widget> pages = <Widget>[
   const MemberPage(),
   RoomListPage(),
-  RegisterPage(),
+  LoginPage(),
   HistoryPage()
 ];
 /* 페이지 실행시 2~4번째 지우고 본인 페이지 넣어주세요 */
